+++ ultracache/purgers.py
@@ -3,56 +3,30 @@ import requests
 from django.conf import settings
 
 
-def varnish(path):
+def broadcast(path, headers=None):
+    # The preferred methody requires RabbitMQ and celery being installed and
+    # configured.
+    from ultracache.tasks import broadcast_purge
+    broadcast_purge.delay(path, headers)
+
+
+def varnish(path, headers=None):
     # See https://www.varnish-software.com/static/book/Cache_invalidation.html
     loc = settings.ULTRACACHE["purge"]["method"]["url"].rstrip("/") + "/" \
         + path.lstrip("/")
     try:
-        r = requests.request("PURGE", loc, timeout=1)
+        r = requests.request("PURGE", loc, timeout=1, headers=headers or {})
     except requests.exceptions.RequestException:
         pass
 
 
-def nginx(path):
+def nginx(path, headers=None):
     # See https://github.com/FRiCKLE/ngx_cache_purge
 
     # Simplest case - one node
     loc = settings.ULTRACACHE["purge"]["method"]["url"].rstrip("/") + "/" \
         + path.lstrip("/")
     try:
-        r = requests.request("PURGE", loc, timeout=1)
+        r = requests.request("PURGE", loc, timeout=1, headers=headers or {})
     except requests.exceptions.RequestException:
         pass
-
-
-def broadcast(path):
-    # The preferred way requires RabbitMQ and celery being installed and
-    # configured.
-    from ultracache.tasks import broadcast_purge
-    broadcast_purge.delay(path)
-
-
-def nginx_more_examples(path):
-    # More nodes
-    for host in ("196.1.2.3", "196.1.2.4"):
-        loc = "http://%s%s" % (host, path)
-        try:
-            r = requests.request("PURGE", loc, timeout=1)
-        except requests.exceptions.RequestException:
-            pass
-
-    # Using RabbitMQ to avoid needing knowledge of the nodes. See
-    # twisted/monitor.py on how to handle the messages. You will need pika
-    # installed.
-    rabbit_host = "192.1.2.5"
-    connection = pika.BlockingConnection(
-        pika.ConnectionParameters(host=rabbit_host)
-    )
-    channel = connection.channel()
-    channel.exchange_declare(exchange="purgatory", type="fanout")
-    channel.basic_publish(
-        exchange="purgatory",
-        routing_key="",
-        body=path
-    )
-    connection.close()