+++ ultracache/tasks.py
@@ -1,9 +1,10 @@
+import json
 import urllib
 import urlparse
 
+from celery import shared_task
 try:
     import pika
-    from celery import shared_task
     DO_TASK = True
 except ImportError:
     DO_TASK = False
@@ -12,9 +13,9 @@ from django.conf import settings
 
 
 @shared_task(max_retries=3, ignore_result=True)
-def broadcast_purge(path):
+def broadcast_purge(path, headers=None):
     if not DO_TASK:
-        raise RuntimeError("Libraries celery and/ or pika not found")
+        raise RuntimeError("Library pika==0.10.0 not found")
 
     try:
         url = settings.ULTRACACHE["rabbitmq-url"]
@@ -35,7 +36,7 @@ def broadcast_purge(path):
     channel.basic_publish(
         exchange="purgatory",
         routing_key="",
-        body=path
+        body=json.dumps({"path": path, "headers": headers or {}})
     )
     connection.close()
     return True