+++ ultracache/tests/test_drf.py
@@ -1,11 +1,11 @@
 import copy
 import json
 
-from django.core.urlresolvers import reverse
+from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.test.utils import override_settings
-from django.conf import settings
+from django.urls import reverse
 
 from rest_framework.test import APIRequestFactory, APIClient
 
@@ -64,25 +64,25 @@ class DRFTestCase(TestCase):
 
     def test_get_dummymodels(self):
         response = self.client.get("/api/dummies/")
-        as_json_1 = json.loads(response.content)
+        as_json_1 = response.json()
 
         # Drop to low level API so post_save does not trigger, meaning the
         # cached version is fetched on the next request.
         DummyModel.objects.filter(pk=self.one.pk).update(title="Onae")
         response = self.client.get("/api/dummies/")
-        as_json_2 = json.loads(response.content)
+        as_json_2 = response.json()
         self.assertEqual(as_json_1, as_json_2)
 
         # Modify it the normal way, which removes the item from cache.
         self.one.title = "Onbe"
         self.one.save()
         response = self.client.get("/api/dummies/")
-        as_json_3 = json.loads(response.content)
+        as_json_3 = response.json()
         self.assertNotEqual(as_json_1, as_json_3)
 
         # Trivial fetch to prove it is cached now
         response = self.client.get("/api/dummies/")
-        as_json_4 = json.loads(response.content)
+        as_json_4 = response.json()
         self.assertEqual(as_json_3, as_json_4)
 
         # Modify via API to confirm that post_save is fired implicitly
@@ -91,7 +91,7 @@ class DRFTestCase(TestCase):
         }
         response = self.client.patch("/api/dummies/%s/" % self.one.pk, data)
         response = self.client.get("/api/dummies/")
-        as_json_5 = json.loads(response.content)
+        as_json_5 = response.json()
         self.assertNotEqual(as_json_4, as_json_5)
 
         # Add an evaluate parameter which leads to a cache miss. Also
@@ -101,12 +101,12 @@ class DRFTestCase(TestCase):
         di["drf"] = {"viewsets": {"*": {"evaluate": "request.user.is_anonymous"}}}
         with override_settings(ULTRACACHE=di):
             response = self.client.get("/api/dummies/")
-            as_json_6 = json.loads(response.content)
+            as_json_6 = response.json()
             self.assertNotEqual(as_json_5, as_json_6)
 
         # Trivial fetch to refresh the cache
         response = self.client.get("/api/dummies/")
-        as_json_7 = json.loads(response.content)
+        as_json_7 = response.json()
 
         # Disable viewset caching which leads to a cache miss. Also
         # surreptiously edit the object so we can confirm a cache miss.
@@ -115,12 +115,12 @@ class DRFTestCase(TestCase):
         di["drf"] = {"viewsets": {object: {}}}
         with override_settings(ULTRACACHE=di):
             response = self.client.get("/api/dummies/")
-            as_json_8 = json.loads(response.content)
+            as_json_8 = response.json()
             self.assertNotEqual(as_json_7, as_json_8)
 
         # Trivial fetch to refresh the cache
         response = self.client.get("/api/dummies/")
-        as_json_9 = json.loads(response.content)
+        as_json_9 = response.json()
 
         # Enable viewset caching for a single viewset class. Also
         # surreptiously edit the object so we can confirm a cache hit.
@@ -129,24 +129,24 @@ class DRFTestCase(TestCase):
         di["drf"] = {"viewsets": {DummyViewSet: {}}}
         with override_settings(ULTRACACHE=di):
             response = self.client.get("/api/dummies/")
-            as_json_10 = json.loads(response.content)
+            as_json_10 = response.json()
             self.assertEqual(as_json_9, as_json_10)
 
     def test_get_dummymodel(self):
         url = "/api/dummies/%s/" % self.one.pk
         response = self.client.get(url)
-        as_json_1 = json.loads(response.content)
+        as_json_1 = response.json()
 
         # Drop to low level API so post_save does not trigger, meaning the
         # cached version is fetched on the next request.
         DummyModel.objects.filter(pk=self.one.pk).update(title="Onfe")
         response = self.client.get(url)
-        as_json_2 = json.loads(response.content)
+        as_json_2 = response.json()
         self.assertEqual(as_json_1, as_json_2)
 
         # Modify it the normal way, which removes the item from cache.
         self.one.title = "Onge"
         self.one.save()
         response = self.client.get(url)
-        as_json_3 = json.loads(response.content)
+        as_json_3 = response.json()
         self.assertNotEqual(as_json_1, as_json_3)